+++ libraries/test_session_common/src/main/java/androidx/media3/test/session/common/MediaSessionConstants.java
@@ -28,6 +28,8 @@ public class MediaSessionConstants {
   public static final String TEST_ON_VIDEO_SIZE_CHANGED = "onVideoSizeChanged";
   public static final String TEST_ON_TRACKS_CHANGED_VIDEO_TO_AUDIO_TRANSITION =
       "onTracksChanged_videoToAudioTransition";
+  public static final String TEST_SET_SHOW_PLAY_BUTTON_IF_SUPPRESSED_TO_FALSE =
+      "testSetShowPlayButtonIfSuppressedToFalse";
 
   // Bundle keys
   public static final String KEY_AVAILABLE_SESSION_COMMANDS = "availableSessionCommands";